firebase provider component separation
